Improving File Access Via External Applications
Numerous changes were made to make an exhaustive search for the file path of a Uri sent to TEdit via an open request from another application such as a file manager. In the event a file path could not be found or the path points to a file TEdit does not have permission to write to the contents of the Uri will be read and displayed as a new file.