Additional Read Permission Checks
TEdit now performs some additional read permission checks and warns accordingly. Previously when attempting to open a file with TEdit through another application TEdit would check for read/write permission and prompt the user to grant permission should it not have already been granted. In the case where the user denied permission TEdit would exit. This behaviour has been changed, when the user denies permission in this case TEdit will still open and display the contents of the file and prompt the user for read/write permission again when attempting to save the document. Of course, should TEdit find that it cannot read the file it will still simply exit.