Opening Extensionless Documents
The file browser will now display all files that do not have an extension. When opening a file without an extension TEdit will perform a rudimentary check of the contents of the file to determine if it is most likely a text file. If the file fails this check the user will be prompted to verify that they still wish to have TEdit attempt to open the file.