DocumentFile.renameTo(String) Fix
Fix for a potential crash when saving to an SDCard. Android's DocumentFile system here apparently leaves open the possibility for vendors to alter the file extension when saving a file. Previously this checked if the name was altered and attempted to rename the document to the originally requested name. Unfortunately it turns out DocumentFile.renameTo(String) doesn't work on DocumentFiles created through DocumentFile.fromSingleUri(Uri)!